Update bot functionalities and error handling
The update includes various modifications. A 'start' command is introduced that sends a greeting message. Changed the method of bot starting from 'bot.start()' to 'run(bot)'. Also, the group_id field in the User object is converted from number to string type. Improved the error handling when creating an invite link in the handleSupergroupWithoutUsername function by adding an inline keyboard with 'Accept' and 'Deny' buttons. When these buttons are clicked, the handleAcceptAction and handleDenyAction will be triggered, which handle the corresponding state changes in the session data.