Update TypeScript and Grammy dependencies
The versions of TypeScript and Grammy used in the project have been updated. In package.json and pnpm-lock.yaml, TypeScript has been upgraded from version 5.4.3 to 5.4.5, and Grammy has been upgraded from version 1.21.1 to 1.22.4. These updates will incorporate any enhancements or fixes made in the newer versions of the dependencies.