Refactor bot's handling of group addition and migration to environment variable
The bot's code has been refactored to handle when it is added to a new group or when it is promoted to an admin in a group better. Additionally, the Mongoose connection now uses an environment variable for the MongoDB URI. The Group model in MongoDB was updated to have most fields set as not required and an additional date_modified field was added. The handling of group addition or removal now creates or looks up a group in the database respectively. The creation of invite links has also been streamlined and now updates the group in the database with the new invite link.